refactor: move classes that are not configuration classes to their corresponding packages
