+++ src/main/java/com/karol/kindergartenmanagementsystem/config/SecurityConfiguration.java
@@ -1,6 +1,8 @@
 package com.karol.kindergartenmanagementsystem.config;
 
 import com.karol.kindergartenmanagementsystem.filter.JwtAuthenticationFilter;
+import com.karol.kindergartenmanagementsystem.security.AccessDeniedHandlerImpl;
+import com.karol.kindergartenmanagementsystem.security.CustomLogoutHandler;
 import com.karol.kindergartenmanagementsystem.service.UserAccountDetailsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
